Moved all procedural mesh generation from rendering to terraformation
