+++ src/world/terraformation/genesis.rs
@@ -13,10 +13,12 @@ use bracket_noise::prelude::{FastNoise, FractalType, NoiseType};
 use futures_lite::future;
 
 use crate::world::{
-    math,
+    math, mesh,
+    rendering::ChunkFacesOcclusion,
     storage::{
-        chunk::{self, Chunk, ChunkKind},
-        voxel, VoxWorld,
+        chunk::{self, Chunk, ChunkKind, ChunkNeighborhood},
+        voxel::{self, FacesOcclusion, VoxelFace, VoxelVertex},
+        VoxWorld,
     },
 };
 
@@ -518,18 +520,119 @@ fn refresh_chunks(world: &mut VoxWorld, locals: impl Iterator<Item = IVec3>) -> 
     let mut result = vec![];
 
     for local in locals {
-        if world.get(local).is_some() {
-            world.update_neighborhood(local);
+        update_neighborhood(world, local);
 
-            //Faces Occlusion
-            //Faces Merging
-            //Vertices Computation
+        if let Some(chunk) = world.get_mut(local) {
+            chunk.vertices = {
+                let occlusion = faces_occlusion(&chunk.kinds);
+
+                if !occlusion.is_fully_occluded() {
+                    let faces = mesh::merge_faces(&occlusion, &chunk);
+                    generate_vertices(faces)
+                } else {
+                    vec![]
+                }
+            };
 
             result.push(local);
         }
     }
 
     result
+}
+
+/**
+Generates vertices data from a given [`VoxelFace`] list.
+
+All generated indices will be relative to a triangle list.
+
+**Returns** a list of generated [`VoxelVertex`].
+*/
+fn generate_vertices(faces: Vec<VoxelFace>) -> Vec<VoxelVertex> {
+    perf_fn_scope!();
+
+    let mut vertices = vec![];
+
+    for face in faces {
+        let normal = face.side.normal();
+
+        for (i, v) in face.vertices.iter().enumerate() {
+            let base_vertex_idx = mesh::VERTICES_INDICES[face.side as usize][i];
+            let base_vertex: Vec3 = mesh::VERTICES[base_vertex_idx].into();
+            vertices.push(VoxelVertex {
+                position: base_vertex + v.as_vec3(),
+                normal,
+            })
+        }
+    }
+
+    vertices
+}
+
+/**
+Computes the faces occlusion data of the given [`ChunkKind`]
+
+**Returns** computed [`ChunkFacesOcclusion`]
+*/
+fn faces_occlusion(chunk: &ChunkKind) -> ChunkFacesOcclusion {
+    perf_fn_scope!();
+
+    let mut occlusion = ChunkFacesOcclusion::default();
+    for voxel in chunk::voxels() {
+        let mut voxel_faces = FacesOcclusion::default();
+
+        if chunk.get(voxel).is_empty() {
+            voxel_faces.set_all(true);
+        } else {
+            for side in voxel::SIDES {
+                let dir = side.dir();
+                let neighbor_pos = voxel + dir;
+
+                let neighbor_kind = if !chunk::is_within_bounds(neighbor_pos) {
+                    let (_, next_chunk_voxel) = chunk::overlap_voxel(neighbor_pos);
+
+                    match chunk.neighborhood.get(side, next_chunk_voxel) {
+                        Some(k) => k,
+                        None => continue,
+                    }
+                } else {
+                    chunk.get(neighbor_pos)
+                };
+
+                voxel_faces.set(side, !neighbor_kind.is_empty());
+            }
+        }
+
+        occlusion.set(voxel, voxel_faces);
+    }
+
+    occlusion
+}
+
+/**
+Updates the [`ChunkNeighborhood`] of a given chunk local.
+This function updates any neighborhood data needed by chunk.
+
+Currently it only updates kind neighborhood data, but in the future, it may update light and other relevant data.
+*/
+fn update_neighborhood(world: &mut VoxWorld, local: IVec3) {
+    if !world.exists(local) {
+        return;
+    }
+
+    let mut neighborhood = ChunkNeighborhood::default();
+    for side in voxel::SIDES {
+        let dir = side.dir();
+        let neighbor = local + dir;
+
+        if let Some(neighbor_chunk) = world.get(neighbor) {
+            neighborhood.set(side, &neighbor_chunk.kinds);
+        }
+    }
+
+    if let Some(chunk) = world.get_mut(local) {
+        chunk.kinds.neighborhood = neighborhood;
+    }
 }
 
 fn generate_chunk(local: IVec3) -> Chunk {
@@ -725,7 +828,7 @@ mod tests {
 
         assert_eq!(dirty_chunks.len(), super::voxel::SIDE_COUNT + 1);
         assert!(dirty_chunks.contains(&local));
-        assert!(world.get(local).is_some(), "Chunk should be added to world");
+        assert!(world.exists(local), "Chunk should be added to world");
 
         let _ = remove_file(path);
 
@@ -740,7 +843,7 @@ mod tests {
         assert_eq!(dirty_chunks.len(), super::voxel::SIDE_COUNT + 1);
         assert!(dirty_chunks.contains(&local));
         assert!(path.exists(), "Cache file should be created by load_chunk");
-        assert!(world.get(local).is_some(), "Chunk should be added to world");
+        assert!(world.exists(local), "Chunk should be added to world");
 
         let _ = remove_file(path);
     }
@@ -1074,6 +1177,246 @@ mod tests {
                 ChunkCmd::Update((1, 1, 1).into(), vec![]),
                 ChunkCmd::Unload((1, 5, 1).into()),
                 ChunkCmd::Unload((1, 6, 1).into())
+            ]
+        );
+    }
+
+    #[test]
+    fn update_neighborhood() {
+        let mut world = VoxWorld::default();
+
+        let center = (1, 1, 1).into();
+        let mut chunk = Chunk::default();
+        chunk.kinds.set_all(10.into());
+        world.add(center, chunk);
+
+        for side in voxel::SIDES {
+            let dir = side.dir();
+            let pos = center + dir;
+            let mut chunk = Chunk::default();
+            chunk.kinds.set_all((side as u16).into());
+            world.add(pos, chunk);
+        }
+
+        super::update_neighborhood(&mut world, center);
+        let chunk = world.get(center).unwrap();
+
+        for side in voxel::SIDES {
+            match side {
+                voxel::Side::Right => {
+                    for a in 0..chunk::Y_AXIS_SIZE {
+                        for b in 0..chunk::Z_AXIS_SIZE {
+                            assert_eq!(
+                                chunk
+                                    .kinds
+                                    .neighborhood
+                                    .get(side, (0, a as i32, b as i32).into()),
+                                Some((side as u16).into())
+                            );
+                        }
+                    }
+                }
+                voxel::Side::Left => {
+                    for a in 0..chunk::Y_AXIS_SIZE {
+                        for b in 0..chunk::Z_AXIS_SIZE {
+                            assert_eq!(
+                                chunk
+                                    .kinds
+                                    .neighborhood
+                                    .get(side, (chunk::X_END as i32, a as i32, b as i32).into()),
+                                Some((side as u16).into())
+                            );
+                        }
+                    }
+                }
+                voxel::Side::Up => {
+                    for a in 0..chunk::X_AXIS_SIZE {
+                        for b in 0..chunk::Z_AXIS_SIZE {
+                            assert_eq!(
+                                chunk
+                                    .kinds
+                                    .neighborhood
+                                    .get(side, (a as i32, 0, b as i32).into()),
+                                Some((side as u16).into())
+                            );
+                        }
+                    }
+                }
+                voxel::Side::Down => {
+                    for a in 0..chunk::X_AXIS_SIZE {
+                        for b in 0..chunk::Z_AXIS_SIZE {
+                            assert_eq!(
+                                chunk
+                                    .kinds
+                                    .neighborhood
+                                    .get(side, (a as i32, chunk::Y_END as i32, b as i32).into()),
+                                Some((side as u16).into())
+                            );
+                        }
+                    }
+                }
+                voxel::Side::Front => {
+                    for a in 0..chunk::X_AXIS_SIZE {
+                        for b in 0..chunk::Y_AXIS_SIZE {
+                            assert_eq!(
+                                chunk
+                                    .kinds
+                                    .neighborhood
+                                    .get(side, (a as i32, b as i32, 0).into()),
+                                Some((side as u16).into())
+                            );
+                        }
+                    }
+                }
+                voxel::Side::Back => {
+                    for a in 0..chunk::X_AXIS_SIZE {
+                        for b in 0..chunk::Y_AXIS_SIZE {
+                            assert_eq!(
+                                chunk
+                                    .kinds
+                                    .neighborhood
+                                    .get(side, (a as i32, b as i32, chunk::Z_END as i32).into()),
+                                Some((side as u16).into())
+                            );
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    #[test]
+    fn faces_occlusion_occlude_empty_chunk() {
+        // Arrange
+        let chunk = ChunkKind::default();
+
+        // Act
+        let occlusions = super::faces_occlusion(&chunk);
+
+        // Assert
+        assert!(
+            occlusions.iter().all(|a| a.is_fully_occluded()),
+            "A chunk full of empty-kind voxels should be fully occluded"
+        );
+    }
+
+    #[test]
+    fn faces_occlusion() {
+        // Arrange
+        let mut chunk = ChunkKind::default();
+
+        // Top-Bottom occlusion
+        chunk.set((1, 1, 1).into(), 1.into());
+        chunk.set((1, 2, 1).into(), 1.into());
+
+        // Full occluded voxel at (10, 10, 10)
+        chunk.set((10, 10, 10).into(), 1.into());
+        chunk.set((9, 10, 10).into(), 1.into());
+        chunk.set((11, 10, 10).into(), 1.into());
+        chunk.set((10, 9, 10).into(), 1.into());
+        chunk.set((10, 11, 10).into(), 1.into());
+        chunk.set((10, 10, 9).into(), 1.into());
+        chunk.set((10, 10, 11).into(), 1.into());
+
+        // Act
+        let faces_occlusion = super::faces_occlusion(&chunk);
+
+        // Assert
+        let faces = faces_occlusion.get((1, 2, 1).into());
+
+        assert_eq!(
+            faces,
+            [false, false, false, true, false, false].into(),
+            "Only down face should be occluded by the bottom voxel"
+        );
+
+        let faces = faces_occlusion.get((1, 1, 1).into());
+
+        assert_eq!(
+            faces,
+            [false, false, true, false, false, false].into(),
+            "Only down face should be occluded by the bottom voxel"
+        );
+
+        let faces = faces_occlusion.get((10, 10, 10).into());
+
+        assert_eq!(
+            faces,
+            [true; voxel::SIDE_COUNT].into(),
+            "Voxel fully surrounded by another non-empty voxels should be fully occluded"
+        );
+    }
+
+    #[test]
+    fn faces_occlusion_neighborhood() {
+        let mut world = VoxWorld::default();
+
+        let mut top = Chunk::default();
+        top.kinds.set_all(2.into());
+
+        let mut down = Chunk::default();
+        down.kinds.set_all(3.into());
+
+        let mut center = Chunk::default();
+        center
+            .kinds
+            .set((0, chunk::Y_END as i32, 0).into(), 1.into());
+        center.kinds.set((1, 0, 1).into(), 1.into());
+
+        world.add((0, 1, 0).into(), top);
+        world.add((0, 0, 0).into(), center);
+        world.add((0, -1, 0).into(), down);
+
+        super::update_neighborhood(&mut world, (0, 0, 0).into());
+        let center = world.get((0, 0, 0).into()).unwrap();
+
+        let faces_occlusion = super::faces_occlusion(&center.kinds);
+
+        let faces = faces_occlusion.get((0, chunk::Y_END as i32, 0).into());
+        assert_eq!(faces, [false, false, true, false, false, false].into());
+
+        let faces = faces_occlusion.get((1, 0, 1).into());
+        assert_eq!(faces, [false, false, false, true, false, false].into());
+    }
+
+    #[test]
+    fn generate_vertices() {
+        // Arrange
+        let side = voxel::Side::Up;
+        let faces = vec![VoxelFace {
+            side,
+            vertices: [
+                (0, 0, 0).into(),
+                (0, 0, 1).into(),
+                (1, 0, 1).into(),
+                (1, 0, 0).into(),
+            ],
+        }];
+
+        // Act
+        let vertices = super::generate_vertices(faces);
+
+        // Assert
+        let normal = side.normal();
+        assert_eq!(
+            vertices,
+            vec![
+                VoxelVertex {
+                    normal,
+                    position: (0.0, 1.0, 0.0).into(),
+                },
+                VoxelVertex {
+                    normal,
+                    position: (0.0, 1.0, 2.0).into(),
+                },
+                VoxelVertex {
+                    normal,
+                    position: (2.0, 1.0, 2.0).into(),
+                },
+                VoxelVertex {
+                    normal,
+                    position: (2.0, 1.0, 0.0).into(),
+                },
             ]
         );
     }
